Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,8 +3,8 @@
 int main() {
     int num1,num2,num3;
     scanf("%d %d %d",&num1,&num2,&num3);
-    int avg = num1+num2+num3/3;
-    printf("Average: %d",avg);
+    float avg = num1+num2+num3/3;
+    printf("Average: %.2f",avg);
 
     
     return 0;
